Handle is empty at

+++ app/src/main/java/uk/co/amlcurran/social/CalendarItem.java
@@ -3,7 +3,7 @@ package uk.co.amlcurran.social;
 public interface CalendarItem {
     String title();
 
-    long startDay();
+    int startDay();
 
     boolean isEmpty();
 }
